requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.51.0
gtts==2.4.0
requests==2.31.0
pydub==0.25.1
httpx==0.27.0